backend-geoloc/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.4
geopandas==1.0.1
h11==0.14.0
idna==3.10
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
xmltodict==0.14.2
